+++ lib/async/job/adapter/active_job/environment.rb
@@ -37,6 +37,12 @@ module Async
 						end
 					end
 					
+					# Number of instances to start. By default (when nil), uses `Etc.nprocessors`.
+					# @returns [Integer | nil]
+					def count
+						nil
+					end
+					
 					# Options to use when creating the container.
 					def container_options
 						{